+++ src/spring/azext_spring/_params.py
@@ -1291,3 +1291,12 @@ def load_arguments(self, _):
         c.argument('service', service_name_type)
         c.argument('execution', help='the name of the job execution.')
         c.argument('job', help='The name of job running in the specified Azure Spring Apps instance.')
+
+    with self.argument_context('spring private-dns-zone add') as c:
+        c.argument('service', service_name_type)
+        c.argument('zone_id', help='The resource id of the private DNS zone which you would like to configure with the service instance.')
+    with self.argument_context('spring private-dns-zone update') as c:
+        c.argument('service', service_name_type)
+        c.argument('zone_id', help='The resource id of the private DNS zone which you would like to configure with the service instance.')
+    with self.argument_context('spring private-dns-zone clean') as c:
+        c.argument('service', service_name_type)