+++ src/aks-preview/HISTORY.rst
@@ -15,6 +15,8 @@ Pending
 4.0.0b4
 ++++++++
 * Add etag support to some aks commands for optimistic concurrency control.
+* Add `--vm-sizes` to `az aks create` and `az aks nodepool add`.
+* Add `az aks nodepool manual-scale add/update/delete` commands.
 
 4.0.0b3
 +++++++
@@ -25,6 +27,8 @@ Pending
 +++++++
 * Improve Windows OutboundNat test case by removing Windows OSSKU limitation
 * `az aks create/update`: add support for new outbound type none
+* Add `az operation show` command to show the details of a specific operation.
+* Add `az operation show-latest` command to show the details of the latest operation.
 
 4.0.0b1
 +++++++